feat(GiniBankSDK): Add Skonto `skonto_switch_tapped` event in RA flow and fix it for just Skonto
PP-926
